% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simTS.R
\name{arma}
\alias{arma}
\title{Create an ARMA model.}
\usage{
arma(ar, ma)
}
\arguments{
\item{ar}{x parameter.}

\item{ma}{Moving average parameter.}
}
\value{
An ARMA model.
}
\description{
Create an ARMA model.
}

