% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecomplexity.R
\name{bspline_err}
\alias{bspline_err}
\title{Function returns result for a single downsample level.}
\usage{
bspline_err(ys, sample_num, max_degree)
}
\arguments{
\item{ys}{A vector or time series.}

\item{sample_num}{The amount the series is downsampled.}

\item{max_degree}{The maximum degree spline polynomial to fit.}
}
\description{
Function returns result for a single downsample level.
}

