% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecomp-methods.R
\name{coef.ecomplex}
\alias{coef.ecomplex}
\title{Return ecomplex coefficients}
\usage{
\method{coef}{ecomplex}(fit)
}
\arguments{
\item{fit}{An object of type 'ecomplex'.}
}
\value{
A dataframe with the two coefficients.
}
\description{
Returns a dataframe with a column for each
coefficient from an ecomplex object.
}

