% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{depth}
\alias{depth}
\title{Check the depth of a list.}
\usage{
depth(x)
}
\arguments{
\item{this}{A list.}
}
\value{
The depth of the list.
}
\description{
Recursively checks the maximum depth of a list.
}

