% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palarm.R
\name{diagn}
\alias{diagn}
\title{Checks estimated global change points for errors 
 and performs update.}
\usage{
diagn(x, kin, delta, m, thresh)
}
\arguments{
\item{x}{Input data.}

\item{kin}{A vector containing estimated global change points.}

\item{delta}{Detect parameter.}

\item{m}{Minimum interval size.}

\item{thresh}{Minimum statistic for change point characterization.
(}
}
\value{
kout Estimated global change points.
}
\description{
Checks estimated global change points for errors 
 and performs update.
}

