% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{downsample_perm}
\alias{downsample_perm}
\title{Create list of all possible index patterns 
 for a given downsampling amount.}
\usage{
downsample_perm(n, ds)
}
\arguments{
\item{n}{Length of list.}

\item{ds}{Downsample rate.}
}
\value{
List of indices
}
\description{
Creates a list of indices that correspond to 
 possible downsampling of a vector of length n
 at the given downsample rate.
}

