% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecomplex.R
\name{ecomplex}
\alias{ecomplex}
\title{Compute the epsilon-complexity of a time series.}
\usage{
ecomplex(x, method = c("cspline", "bspline", "lift", "all"), ds = 6,
  max_degree = 5)
}
\arguments{
\item{x}{A vector of points.}

\item{method}{The interpolation or approximation method. One of
c("bspline", "cspline")}

\item{ds}{Number of times to downsample the input sequence.}

\item{max_degree}{The maximum order spline used in the approximation
step}
}
\value{
A list of features or the slope coefficient of the fit.
}
\description{
Compute the epsilon-complexity of a time series.
}

