% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecomplexity.R
\name{ecomplexity}
\alias{ecomplexity}
\title{Compute the epsilon-complexity of a time series.}
\usage{
ecomplexity(xx, ds = 5, method = c("cspline", "bspline", "adlift"),
  max_degree = 4)
}
\arguments{
\item{ds}{Number of times to downsample.}

\item{ys}{A sequence of points.}
}
\value{
A list of features or the slope coefficient of the fit
}
\description{
Compute the epsilon-complexity of a time series.
}

