% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/features.R
\name{eeg_features}
\alias{eeg_features}
\title{Extract eeg features}
\usage{
eeg_features(x, features = NULL, parallel = TRUE)
}
\arguments{
\item{x}{The number of time series generated for each group}

\item{features}{The features set to run}

\item{parallel}{Parallelize computations (Linux only)}
}
\value{
A data frame with features as columns and an id
 column identifying group membership
}
\description{
Extract eeg features
}

