% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/features.R
\name{extract_features}
\alias{extract_features}
\title{Extract set of features from a time series.}
\usage{
extract_features(data, features, id = "1", para = TRUE, verbose = FALSE,
  ...)
}
\arguments{
\item{data}{A time series, matrix of dataframe.}

\item{features}{A list or vector of ts_features.}

\item{prefix}{Optional prefix for column headers.}
}
\value{
A data frame with row of features for each time.
 series in data. (TODO: single feature, system 
 check for parallelization)
}
\description{
Extract set of features from a time series.
}

