% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simTS.R
\name{farima_mod}
\alias{farima_mod}
\title{Create a FARIMA model.}
\usage{
farima_mod(ar, ma, d)
}
\arguments{
\item{ar}{The autoregressive parameters.}

\item{ma}{The moving average parameters.}

\item{d}{The long term memory parameter.}
}
\value{
A function model of class "farima".
}
\description{
Create a FARIMA model.
}

