% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palarm.R
\name{finalal}
\alias{finalal}
\title{Calculate final change point estimates and mean of each interval
 Based on Matlab function of Anatoly Zlotnik and Alexandra Piryatinska}
\usage{
finalal(X, xin, kin, m, delta, thresh)
}
\arguments{
\item{X}{Input data.}

\item{xin}{The original time-series.}

\item{kin}{A vector containing estimated global change points}

\item{m}{minimum interval size.}

\item{delta}{detect parameter

A \code{list} with :

\tabular{ll}{
\code{meanK} \tab the mean values of the intervals,\cr
\code{kout} \tab the final estimated change points.\cr
}}
}
\description{
Calculate final change point estimates and mean of each interval
 Based on Matlab function of Anatoly Zlotnik and Alexandra Piryatinska
}

