% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simTS.R
\name{fsuite}
\alias{fsuite}
\title{Generate A Default suite of function models.}
\usage{
fsuite()
}
\value{
A list of two lists of function models with
         the same functions and varying default parameters.
}
\description{
Generate A Default suite of function models.
}

