% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simTS.R
\name{generate}
\alias{generate}
\title{Generate a function from a model}
\usage{
generate(mod)
}
\arguments{
\item{mod}{A function model}
}
\value{
Returns a function based on the model 
 parameters in mod
}
\description{
Generate a function from a model
}

