% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/features.R
\name{hurst}
\alias{hurst}
\title{Corrected Hurst exponent}
\usage{
hurst(x)
}
\arguments{
\item{x}{The data}
}
\value{
The features
}
\description{
The pracma implementation of the corrected Hurst
 exponent
}

