% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lift.R
\name{lift_comp}
\alias{lift_comp}
\title{Compute epsilon-complexity using the lifting scheme}
\usage{
lift_comp(xx)
}
\arguments{
\item{xx}{A time series or vector}
}
\value{
A \code{list} with :

\tabular{ll}{
\code{coefficients} \tab The epsilon-complexity coefficients,\cr
\code{fit} \tab The complete linear model generated by lm(), \cr
\code{epsilons} \tab The mean sum of absolute errors at each level \cr
\code{S}     \tab The fraction of points removed at each level 
}
}
\description{
Computes the epsilon-complexity coefficient of a 
  time series using a wavelet lifting scheme.
}

