% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simTS.R
\name{logistic}
\alias{logistic}
\title{Create a logistic function model.}
\usage{
logistic(r)
}
\arguments{
\item{r}{Model parameter.}
}
\value{
A logistic function model.
}
\description{
Create a logistic function model.
}

