% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simTS.R
\name{mackeyglass_mod}
\alias{mackeyglass_mod}
\title{Create a Mackey-Glass model}
\usage{
mackeyglass_mod(tau, init)
}
\arguments{
\item{tau}{The time lag}

\item{init}{A vector used to initialize the Mackey-Glass function}
}
\value{
A model for a Mackey-Glass equation
}
\description{
Create a Mackey-Glass model
}

