% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neville.R
\name{neville}
\alias{neville}
\title{Evaluate the Lagrange polynomials at position.
 
Compute the value of the Lagrange interpolating polynomial at a 
 position xs. The function is borrowed from the \code{pracma} package.}
\usage{
neville(x, y, xs)
}
\arguments{
\item{x}{The x locations of the function values y = f(x).}

\item{y}{The funtion values f(x) of the function to be interpolated.}

\item{xs}{The point at which to evaluate the lagrange polynomial.}
}
\value{
Interpolated value.
}
\description{
Evaluate the Lagrange polynomials at position.
 
Compute the value of the Lagrange interpolating polynomial at a 
 position xs. The function is borrowed from the \code{pracma} package.
}

