% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{normalize}
\alias{normalize}
\title{Normalize sequence values to (0,1).}
\usage{
normalize(x)
}
\arguments{
\item{x}{The sequence to normalize.}
}
\value{
The normalized sequence.
}
\description{
Computes (x - min x)/(max x - min x).
}

