% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nterm_complexity.R
\name{nterm}
\alias{nterm}
\title{Find Nterm Complexity of a Time Series.}
\usage{
nterm(ys, verbose = FALSE, slope_only = FALSE, ...)
}
\arguments{
\item{ys}{A vector, time series.}

\item{verbose}{Print debugging information.}

\item{slope_only}{Return only the slope.}
}
\value{
A list or the slope coefficient of the nterm approximation
}
\description{
Estimate of the complexity of a function based 
on the rate of error decay as fewer terms are used
to approximate the function.
}

