% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nterm_complexity.R
\name{nterm3}
\alias{nterm3}
\title{Find Nterm Complexity of a Time Series.}
\usage{
nterm3(ys, verbose = FALSE, ...)
}
\arguments{
\item{ys}{A vector, time series (or matrix ... ?)}

\item{verbose}{Print debugging information}
}
\value{
A list or the slope coefficient of the nterm approximation
}
\description{
Find the nterm complexity of a function
}

