% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palarm.R
\name{palarm}
\alias{palarm}
\title{Estimation of change points and mean on each
stable interval.}
\usage{
palarm(x, delta1 = 1, delta2 = 0, pf = 0.1, m = 5, M = 1000,
  epsilon = 0.02)
}
\arguments{
\item{x}{Input data.}

\item{delta1}{A weighting parameter for phase I.}

\item{delta2}{A weighting parmameter for phase II.}

\item{pf}{False detection probability.}

\item{m}{Minimum interval size.}

\item{M}{Number of terms in KS series.}

\item{epsilon}{Minimum relative distance of change point 
from endpoints of local interval.}
}
\value{
A \code{list} with :

\tabular{ll}{
\code{kout} \tab Vector of indices of the detected change points,\cr
\code{means} \tab Vector of mean values on each stationary segment. \cr
}
}
\description{
Estimation of change points and mean on each
stable interval.
}

