% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecomp-methods.R
\name{plot.ecomplex}
\alias{plot.ecomplex}
\title{Plot an ecomplex object}
\usage{
\method{plot}{ecomplex}(x, ...)
}
\arguments{
\item{x}{The list returned by callling ecomplex()}

\item{...}{Additional arguments to base plot function.}
}
\description{
Plot the linear fit log(epsilons) ~ log(S) 
 where epsilons are the errors at each downsample level
 and S is the fraction of points retained at a given
 downsample level.
}

