% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simTS.R
\name{quadratic_mod}
\alias{quadratic_mod}
\title{Create a quadratic function model.}
\usage{
quadratic_mod(x0)
}
\arguments{
\item{x0}{Initial value (TODO: ).}
}
\value{
A quadratic function model.
}
\description{
Create a quadratic function model.
}

