% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{random_sample}
\alias{random_sample}
\title{Create indices to randomly downsample a sequence.}
\usage{
random_sample(n, ds)
}
\arguments{
\item{n}{Length of list.}

\item{ds}{Downsample rate.}
}
\value{
List of indices
}
\description{
Creates a list of indices that sample a fractional
 of the total values. The fraction sampled is
 1/2:ds, the same fraction sampled in downsample_perm.
}

