% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{size}
\alias{size}
\title{Return 2-D Size of Input.}
\usage{
size(x)
}
\arguments{
\item{x}{Any one or two dimensional R data structure.}
}
\value{
A numeric vector of length two: c(#rows, #cols).
}
\description{
Matlab style size() function for vectors, lists, 
matrices or data.frames.
}

