% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/features.R
\name{test_features}
\alias{test_features}
\title{Test a default suite of time series features.}
\usage{
test_features(n = 10, seed = 2017, features = NULL, para = FALSE)
}
\arguments{
\item{n}{The number of time series generated for each group.}

\item{seed}{The seed for the random number generator.}
}
\value{
A data frame with features as columns and an id
 column identifying group membership.
}
\description{
Extracts default set of time series features from
 two groups of ARMA(2,2) processes. ARMA processes 
 are generated using arima.sim() using different initial
 parameters for each group.
}

