% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simTS.R
\name{weierstrass}
\alias{weierstrass}
\title{Create an Weierstrass function model.}
\usage{
weierstrass(a, b, random = TRUE)
}
\arguments{
\item{a}{The amplitudes parameter.}

\item{b}{The frequencies parameter.}

\item{random}{Create a random phase model.}
}
\value{
The parametrized model.
}
\description{
Creates a model for a Weierstrass or random 
 phase Weierstrass function
}

