% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simTS.R
\name{weierstrass_a}
\alias{weierstrass_a}
\title{Single parameter Weierstrass function model.}
\usage{
weierstrass_a(a, random = TRUE)
}
\arguments{
\item{a}{The holder constant of the function. Equal to 
-log(a)/log(b) for the two parameter Weierstrass}

\item{random}{If true generates a random phase model.}
}
\value{
The parametrized model.
}
\description{
Creates a model for a Weierstrass or random 
 phase Weierstrass function with single parameter
 a. The resulting function is a-Holder continuous.
 If random, theu function is evaluated on (1,n), 
 otherwise on (0,1),
}

