% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palarm.R
\name{ystat}
\alias{ystat}
\title{Calculate statistics for change point and
 maximize to detect change point k.}
\usage{
ystat(X, delta, m)
}
\arguments{
\item{X}{Input data.}

\item{delta}{Detect parameter.}

\item{m}{Minimum interval size.}
}
\value{
A \code{list} with :

\tabular{ll}{
\code{ystat} \tab The test stastistic,\cr
\code{k} \tab The estimated change point \cr
}
}
\description{
Calculate statistics for change point and
 maximize to detect change point k.
}

