% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecomplex.R
\name{cspline_err}
\alias{cspline_err}
\title{Function returns result for a single downsample level.}
\usage{
cspline_err(y, sample_num, max_degree = NULL)
}
\arguments{
\item{y}{A vector or time series.}

\item{sample_num}{The amount the series is downsampled.}

\item{max_degree}{The maximum degree spline polynomial to fit.}
}
\description{
Function returns result for a single downsample level.
}

