% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cspline.R
\name{cspline_err}
\alias{cspline_err}
\title{Function returns result for a single downsample level.}
\usage{
cspline_err(y, sample_num, max_degree, err_norm, sample_type)
}
\arguments{
\item{y}{A vector or time series.}

\item{sample_num}{The amount the series is downsampled.}

\item{max_degree}{The maximum degree spline polynomial to fit.}

\item{err_norm}{The norm used in computing the approximation error.}

\item{sample_type}{Downsampling pattern to use.}
}
\value{
The mean errors for given sample_num
}
\description{
Function returns result for a single downsample level.
}

