% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecomplex.R
\name{ecomplex}
\alias{ecomplex}
\title{Compute the epsilon-complexity of a time series.}
\usage{
ecomplex(x, ds = 6, max_degree = 5, method = c("cspline", "bspline",
  "lift", "all"), err_norm = c("mae", "mse", "max"), sample_type = c("step",
  "random"))
}
\arguments{
\item{x}{A vector of points.}

\item{ds}{Number of times to downsample the input sequence.}

\item{max_degree}{The maximum order spline used in the approximation
step}

\item{method}{The interpolation or approximation method. One of
c("bspline", "cspline")}

\item{err_norm}{The norm type used in computing the approximation error.}

\item{sample_type}{The downsampling type. Either randomly sampled
or downsampled in integer steps.}
}
\value{
A \code{list} with :
\tabular{ll}{
\code{A}        \tab The epsilon-complexity intercept coefficient \cr
\code{B}        \tab The epsilon-complexity slope coefficient \cr
\code{fit}      \tab The full linear model generated by fitting log(epsilons) ~ log(S) using \code{lm()}. \cr
\code{epsilons} \tab The mean sum of absolute errors at each downsample level. \cr
\code{S}        \tab The fraction of samples maintained at each downsample level. \cr
\code{method}  \tab The method used or a list of methods if method 
                     "all" is used. 
}
}
\description{
Compute the epsilon-complexity of a time series.
}

