% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecomplex.R
\name{get_epsilons}
\alias{get_epsilons}
\title{Generate a function from a model}
\usage{
get_epsilons(mod)
}
\arguments{
\item{mod}{A function model}
}
\value{
Returns a function based on the model 
 parameters in mod
}
\description{
Generate a function from a model
}

