% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecomplexity.R
\name{normalize}
\alias{normalize}
\title{Normalize sequence values to (0,1)}
\usage{
normalize(xx)
}
\arguments{
\item{xx}{The sequence to normalize.}
}
\value{
The normalized sequence.
}
\description{
Normalize sequence values to (0,1)
}

